Cid/front

+++ src/Components/Cart/CartProduct.js
@@ -8,28 +8,33 @@ import { changeQtd } from "../../Services/api";
 
 
 export default function CartProduct({ product }) {
-    
-    const {renderCart, setRenderCart} = useContext(GlobalContext);
+
+    const { renderCart, setRenderCart } = useContext(GlobalContext)
     const [qtd, setQtd] = useState(product.qtd);
     const token = localStorage.getItem("token");
 
-    async function deleteItem() {
-        try {
-            await deleteProduct(product._id, getConfig(token));
-            alert("item excluído com sucesso!");
-            setRenderCart(!renderCart);
-        } catch (error) {
-            
+    function deleteItem() {
+
+        if (window.confirm('Tem certeza que deseja deletar o item?')) {
+            const promisse = deleteProduct(product._id, getConfig(token))
+
+            promisse
+                .catch(err => {
+                    console.log(err)
+                })
+                .then(res => {
+                    setRenderCart(!renderCart)
+                })
         }
     }
-    
-    async function handleInput(e) {
+
+    function handleInput(e) {
         setQtd(Number(e.target.value));
-        const body = { productId: product._id, newQtd: Number(e.target.value) }; 
-        changeQtd (body, getConfig(token));
+        const body = { productId: product._id, newQtd: Number(e.target.value) };
+        changeQtd(body, getConfig(token));
         setRenderCart(!renderCart);
     }
-    
+
     return (
 
         <ProductWrapper>
@@ -37,7 +42,7 @@ export default function CartProduct({ product }) {
             <TextWrapper>
                 <h1>{product.name}</h1>
                 <h2>{product.artist}</h2>
-                <h3>{(Number(product.price.replace(",", "."))).toLocaleString('pt-BR', {style:"currency", currency:"BRL"})}</h3>
+                <h3>{(Number(product.price.replace(",", "."))).toLocaleString('pt-BR', { style: "currency", currency: "BRL" })}</h3>
             </TextWrapper>
             <Funcs>
                 <AddDelete>
@@ -46,52 +51,54 @@ export default function CartProduct({ product }) {
                         type='number'
                         value={qtd}
                         min="1" />
-                    <p onClick={deleteItem}>
+
+                    <h1 onClick={deleteItem}>
                         Delete
-                    </p>
+                    </h1>
 
                 </AddDelete>
                 <PriceWrapper>
-                    {(Number(product.price.replace(",", "."))*qtd).toLocaleString('pt-BR', {style:"currency", currency:"BRL"})}
+                    {(Number(product.price.replace(",", ".")) * qtd).toLocaleString('pt-BR', { style: "currency", currency: "BRL" })}
                 </PriceWrapper>
             </Funcs>
         </ProductWrapper>
-
-
-
-
     )
 }
 
 const ProductWrapper = styled.div`
 width: 70vw;
-height: 160px;
+height: 140px;
 margin-top: 10px;
 padding: 10px;
 
-background-color: aliceblue;
+background-color: #caf0f8;
 
 display: flex;
 justify-content: flex-start;
 
 img{
+    width: 120px;
+    height: 120px;
+    max-width: 150px;
     margin-right: 15px;
 }
 `
 
 const TextWrapper = styled.div`
-width: 30vw;
+width: 40vw;
 height: 100%;
 
-background-color: aliceblue;
+background-color: #caf0f8;
 
 h1{
     margin-bottom: 10px;
+    color:#03045e;
 }
 
 h2{
     font-size: 12px;
     color: grey;
+    margin-bottom: 10px;
 
 }
 `
@@ -100,7 +107,7 @@ const Funcs = styled.div`
 width: 100%;
 height: 100%;
 
-background-color: lightsalmon;
+background-color: #caf0f8;
 
 display: flex;
 justify-content: flex-end;
@@ -115,7 +122,12 @@ flex-direction: column;
 align-items: center;
 justify-content: center;
 
-background-color:green /* aliceblue */;
+background-color:#caf0f8 ;
+
+h1{
+    color: #03045e;
+    cursor:pointer;
+}
 
 input{
     width: 50px;
@@ -133,6 +145,6 @@ margin-left: 10px;
 display: flex;
 justify-content: center;
 align-items: center;
-background-color: lightblue;
+background-color: #caf0f8;
 
 `
